fixed link hovering, added more logging to widgets
It turns out that the reason the links weren't showing up right was because I didn't tab the incrementing out right in Client's `_parse_gmi` function. 
As it stood it was only being incremented when it got to a line that didn't actually have a link in it, which looked very strange when cycling through links in the app.
Fixes #7

Modified:
networking/client.py
   - Moved the idx incrementing outside the if/else statements in the for line in lines loop

ui/view.py, ui/widgets/gem_page.py
   - Added logging to more inner functions