Relative links now work, tabs after link url formatting fixed
Relative links now work as they should in the browser. LagannApp now takes the URL from the current `center_page` (really needs to be renamed), and appends the desired url to the end of it. 
It then passes this to the client through `get_page()`. 

There was a bug where links where not being formatted correctly due to multiple tabs being used to separate the URL the link leads to from the placeholder text. Such as `b"=> "gemini://google.com/\t\tThis is the placeholder text". 
This text would have been formatted as `" This is placeholder text"`, with an awkward space in the beginning. This was fixed by modifying the parsing rules in `_gmi_parse_line()` from networking/client.py 

Modified:
networking/client.py
   - Added better parsing rules for links, now takes into account multiple tabs after the desired link destination

networking/lagcerts.py
   - Removed a line of logging from `valid_cert()`, not really necessary to log that portion since that's the best case scenario

ui/view.py
   - View now appends the end of the desired URL to the current URL before passing it to the client, if it's a relative link.