SNI, relative links, MIME parsing, and preformatted lines now working
Fixed SNI issues (needed to specify hostname when wrapping socket), and problems with my MIME parsing
Added support for preformatted lines, and full support for relative links 

Added testing to the project! Making sure the parsing worked as intended without utilizing automated testing would've been a huge pain in the derriere. Now knowing how to set this up, this opens up a lot more assurances through testing. 

`_parse_query()` now correctly formats relative links! This function really should be renamed to `_format_query()`, since it's not really parsing.

Closes #11

Moved `_parse_query()` from connection.py to client.py . The point of the connection is to represent a 'connection', not really for it to be doing any other kinds of logic other than providing a further abstraction over socket operations. Because of this I moved `_parse_query` to client and made the necessary changes. 

Fixed a bug in `_parse_MIME()` that resulted in the app crashing if the server didn't provide a charset. 
The issue was that I had assumed that the headers would have a charset in place if they were of type `text/gemini`. Fixed by adding an if statement that checks for the lack of a charset property and sets it to `utf-8`.


Modified parameters for `get_page()`, added `relative = None`. This parameter will only be set if the link is relative, and will contain the  relative URL (e.g. `../testing/something.gmi`)

Modified return type of `get_page()`. Now returns a tuple containing a `DisplayForm` and the query. It now returns the query so that the UI can set the search bar text to it. Since that was not previously being done when navigating through links on the page.

Modified operations in `LagannApp` to reflect the new changes made to `get_page()`.

Modified `gem_page.py`, added more debug logging statements to it.


 


Created:
ui/__init__.py
   - For indexing

networking/test_networking.py
   - Added testing for `_parse_query()` (may be renamed `_format_query`)

Modified:
networking/client.py
   - Moved `_parse_query()` here from `connection.py`
   - Added support for relative links in `_parse_query()`
   - Added support for preformatted lines in `_gmi_parse_line()`
   - Fixed a bug in `_parse_MIME()` with charsets not being set if they were absent

networking/connection.py
   - Moved `_parse_query()` to `client.py`
      + Modified operations to reflect the change

ui/view.py
   - Modified operations to reflect the changes that were made in `client.py`
   - Added more debug logging statements

ui/widgets/gem_page.py
   - Added more debug logging statements